move sse header extract function to a different file
