+++ pkg/block/s3/adapter.go
@@ -526,77 +526,6 @@ func (a *Adapter) copyPart(ctx context.Context, sourceObj, destinationObj block.
 	}, nil
 }
 
-// extractAmzServerSideHeaders extracts the x-amz-server-side-* headers from the response
-// TODO(barak): verify that this is the correct way to extract the headers
-func extractSSHeaderUploadPartCopy(resp *s3.UploadPartCopyOutput) http.Header {
-	// x-amz-server-side-* headers
-	headers := make(http.Header)
-	if resp.SSECustomerAlgorithm != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Customer-Algorithm", *resp.SSECustomerAlgorithm)
-	}
-	if resp.SSECustomerKeyMD5 != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Customer-Key-Md5", *resp.SSECustomerKeyMD5)
-	}
-	if resp.SSEKMSKeyId != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Aws-Kms-Key-Id", *resp.SSEKMSKeyId)
-	}
-	if resp.ServerSideEncryption != "" {
-		headers.Set("X-Amz-Server-Side-Encryption", string(resp.ServerSideEncryption))
-	}
-	return headers
-}
-
-func extractSSHeaderCreateMultipartUpload(resp *s3.CreateMultipartUploadOutput) http.Header {
-	// x-amz-server-side-* headers
-	headers := make(http.Header)
-	if resp.SSECustomerAlgorithm != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Customer-Algorithm", *resp.SSECustomerAlgorithm)
-	}
-	if resp.SSECustomerKeyMD5 != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Customer-Key-Md5", *resp.SSECustomerKeyMD5)
-	}
-	if resp.SSEKMSKeyId != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Aws-Kms-Key-Id", *resp.SSEKMSKeyId)
-	}
-	if resp.ServerSideEncryption != "" {
-		headers.Set("X-Amz-Server-Side-Encryption", string(resp.ServerSideEncryption))
-	}
-	if resp.SSEKMSEncryptionContext != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Context", *resp.SSEKMSEncryptionContext)
-	}
-	return headers
-}
-
-func extractSSHeaderCompleteMultipartUpload(resp *s3.CompleteMultipartUploadOutput) http.Header {
-	// x-amz-server-side-* headers
-	headers := make(http.Header)
-	if resp.SSEKMSKeyId != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Aws-Kms-Key-Id", *resp.SSEKMSKeyId)
-	}
-	if resp.ServerSideEncryption != "" {
-		headers.Set("X-Amz-Server-Side-Encryption", string(resp.ServerSideEncryption))
-	}
-	return headers
-}
-
-func extractSSHeaderUploadPart(resp *s3.UploadPartOutput) http.Header {
-	// x-amz-server-side-* headers
-	headers := make(http.Header)
-	if resp.SSECustomerAlgorithm != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Customer-Algorithm", *resp.SSECustomerAlgorithm)
-	}
-	if resp.SSECustomerKeyMD5 != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Customer-Key-Md5", *resp.SSECustomerKeyMD5)
-	}
-	if resp.SSEKMSKeyId != nil {
-		headers.Set("X-Amz-Server-Side-Encryption-Aws-Kms-Key-Id", *resp.SSEKMSKeyId)
-	}
-	if resp.ServerSideEncryption != "" {
-		headers.Set("X-Amz-Server-Side-Encryption", string(resp.ServerSideEncryption))
-	}
-	return headers
-}
-
 func (a *Adapter) UploadCopyPart(ctx context.Context, sourceObj, destinationObj block.ObjectPointer, uploadID string, partNumber int) (*block.UploadPartResponse, error) {
 	var err error
 	defer reportMetrics("UploadCopyPart", time.Now(), nil, &err)
